% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyOperation.R
\name{applyOperation}
\alias{applyOperation}
\title{Apply an operation on a data.frame}
\usage{
applyOperation(data, identifier)
}
\arguments{
\item{data}{A \code{data.frame}. The operation will be applied on each of its rows.}

\item{identifier}{A \code{character}. The operation will be selected according to it.}
}
\value{
A \code{numeric array} containing the results, if the identifier is valid.
        An error message, otherwise.
}
\description{
Applies a pre-defined operation on each row of a data.frame.
}
\section{Valid identifiers}{

The operation will be selected according to the following identifiers:
\enumerate{
   \item Constant .AVG: \code{\link[base]{mean}}
   \item Constant .STD: \code{\link[stats]{sd}}
   \item Constant .MEDIAN: \code{\link[stats]{median}}
}
}

