% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\title{Connect to the dzVis database}
\usage{
connect()
}
\value{
The \code{DBI} connection object to access the dzVis database
}
\description{
Connects to the dzVis database and provides a connection object to access it.
}
\details{
Detailed information about the dzVis database can be found on
}
\seealso{
\code{\link{disconnect}}
}

