% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createComboChart.R
\name{createComboChart}
\alias{createComboChart}
\title{Create a Google Combo Chart}
\usage{
createComboChart(table, targetVar, groupVar, timeVar, min = NULL,
  max = NULL, restrictions = NULL)
}
\arguments{
\item{table}{The name of the table from which the data should be retrieved}

\item{targetVar}{The numeric value of interest - the vertical axis variable}

\item{groupVar}{Variable by which \code{targetVar} should be grouped}

\item{timeVar}{The horizontal axis variable}

\item{min}{Lower bound for the \code{timeVar}}

\item{max}{Upper bound for the \code{timeVar}}

\item{restrictions}{n x 2 matrix containing the n equality restrictions that make timeVar and groupVar unique when combined.}
}
\value{
\code{TRUE} if the chart was successfully created and
        \code{FALSE} otherwise
}
\description{
Creates a Combo Chart using the googleVis R package
}
\section{Validation of the parameters}{

\enumerate{
   \item Ola hahaha
   \item Hue olaaa
}
}
\seealso{
\code{googleVis}
}

