% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createJSONFile.R
\name{createJSONFile}
\alias{createJSONFile}
\title{Create a JSON file}
\usage{
createJSONFile(data, filename = "output.json")
}
\arguments{
\item{data}{A list object to be converted to a json object}

\item{filename}{The name of the output file}
}
\value{
\code{TRUE} if the file was correctly created and
        \code{FALSE} otherwise
}
\description{
Creates a .JSON file from a list, encoded with UTF-8.
}

