% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugCharacters.R
\name{debugCharacters}
\alias{debugCharacters}
\title{Debug characters}
\usage{
debugCharacters(string)
}
\arguments{
\item{string}{A \code{character}. The string to be debugged}
}
\value{
string A \code{character}. The converted string
}
\description{
Converts dirty and special characters that were misinterpreted by OpenCPU as Latin1 instead of UTF-8
}

