% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disconnect.R
\name{disconnect}
\alias{disconnect}
\title{Disconnect from the dzVis database}
\usage{
disconnect(connection)
}
\arguments{
\item{connection}{The connection object to access the dzVis database}
}
\description{
Disconnect from the dzVis database
}
\details{
Detailed information about the dzVis database can be found on
}
\seealso{
\code{\link{connect}}
}

