% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatDates.R
\name{formatDates}
\alias{formatDates}
\title{Format dates}
\usage{
formatDates(table, dates, toStandard = TRUE, connection = NULL)
}
\arguments{
\item{table}{bla}

\item{dates}{bla}

\item{toStandard}{bla}

\item{connection}{bla}
}
\value{
foadoasas
}
\description{
blblba
}

