% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatGoogleChartTitle.R
\name{formatGoogleChartTitle}
\alias{formatGoogleChartTitle}
\title{Format a Google Chart title}
\usage{
formatGoogleChartTitle(table, restrictions, targetVar)
}
\arguments{
\item{table}{A \code{character}. The name of the table.}

\item{restrictions}{A n x 2 \code{matrix}. The n equality restrictions that make timeVar and groupVar values unique when combined.}

\item{targetVar}{A \code{character}. The name of the column with the vertical axis variable.}

\item{connection}{A \code{DBI} Connection object if a new connection should not be estabilished.}
}
\value{
A \code{character}. The title that was generated.
}
\description{
Generates a meaningful title for a Google Chart, using some parameters that were chosen to create the chart.
}

