% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAttributeValues.R
\name{getAttributeValues}
\alias{getAttributeValues}
\title{Get attribute values from a dzVis table}
\usage{
getAttributeValues(attribute, table, nvalues)
}
\arguments{
\item{attribute}{The name of the column to be selected}

\item{table}{The name of the table to be selected}

\item{nvalues}{The number of values to be selected from the top.}
}
\value{
The name of the file in which the results were written.
}
\description{
Get top values of an attribute - a column - from a dzVis database table
and save the results to a json file.
}
\seealso{
\code{\link{connect}}
}

