% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getColumnAlias.R
\name{getColumnAlias}
\alias{getColumnAlias}
\title{Get column description alias}
\usage{
getColumnAlias(table, column, connection = NULL)
}
\arguments{
\item{table}{A \code{character}. The name of the table that contains the target column.}

\item{column}{A \code{character}. The name of the target column.}

\item{connection}{A \code{DBI} Connection object if a new connection should not be estabilished.}
}
\value{
A \code{character}. The alias of the column description.
}
\description{
Gets the alias of a column description from a dzVis database table.
}

