% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getColumnsByCategory.R
\name{getColumnsByCategory}
\alias{getColumnsByCategory}
\title{Get columns by category}
\usage{
getColumnsByCategory(table, category, connection = NULL)
}
\arguments{
\item{table}{A \code{character}. The name of the table.}

\item{category}{A \code{character}. The name of the category.}

\item{connection}{A \code{DBI} Connection object if a new connection should not be estabilished.}
}
\value{
A \code{character vector} containing the names of the valid columns.
        If the parameter \code{category} is not valid, an error message.
}
\description{
Gets the column names of all columns that satisfy the criteria to represent a Google Chart category.
             This information is retrieved from the corresponding description table in the dzVis database.
}

