% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getColumnValues.R
\name{getColumnValues}
\alias{getColumnValues}
\title{Get attribute values from a dzVis table}
\usage{
getColumnValues(column, table, nvalues)
}
\arguments{
\item{column}{A \code{character} value. The name of the column to be selected}

\item{table}{A \code{character} value. The name of the table to be selected}

\item{nvalues}{A \code{numeric} value. The number of values to be selected from the top.}
}
\value{
The name of the file in which the results were written.
}
\description{
Gets the top values of a column from a dzVis database table and saves the results in a .json file.
}
\seealso{
\code{\link{connect}}
}

