% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataTypes.R
\name{getDataTypes}
\alias{getDataTypes}
\title{Get column types}
\usage{
getDataTypes(table, columns, connection = NULL)
}
\arguments{
\item{table}{A \code{character}. The name of the table.}

\item{columns}{A \code{character vector}. The name(s) of the column(s).}

\item{connection}{A \code{DBI} Connection object if a new connection should not be estabilished.}
}
\value{
A \code{data.frame}. The column \code{cod} contains the columns names. The column \code{ctype}, their types.
}
\description{
Gets the data types of table columns. This information is retrieved from the corresponding description table in the dzVis database.
}

