% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLimits.R
\name{getLimits}
\alias{getLimits}
\title{Get limits of a column}
\usage{
getLimits(table, column, connection = NULL)
}
\arguments{
\item{table}{A \code{character}. The name of the table.}

\item{column}{A \code{character}. The name of the column.}

\item{connection}{A \code{DBI} Connection object if a new connection should not be estabilished.}
}
\value{
A \code{data.frame}. The first column contains the minimum value. The second, the maximum.
}
\description{
Gets the maximum and minimum value of a column from a dzVis database table.
}

