% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{importData}
\alias{importData}
\title{Import data from the dzVis database}
\usage{
importData(table, columns, restrictions = NULL, limits = NULL,
  alternatives = NULL, connection = NULL)
}
\arguments{
\item{table}{A \code{character}. The name of the table.}

\item{columns}{A \code{character vector}. The name(s) of the column(s). If all columns should be imported, this argument must be set to \code{*}.}

\item{restrictions}{A n x 2 \code{matrix}. Equality restrictions. The first column must contain the names of the columns. The second, its values.}

\item{limits}{A n x 3 \code{matrix}. Inequality restrictions. See the details section for important information.}

\item{alternatives}{A n x 2 \code{matrix}. Alternative values of a column (joined by 'or' conditions in a SQL query). The first column must contain the names of the columns. The second, its values.}

\item{connection}{A \code{DBI} Connection object if a new connection should not be estabilished.}
}
\value{
A \code{data.frame} with the requested data.
}
\description{
Gets data from the dzVis database according to the specified table and columns.
             Original column types are set to its equivalents in R.
}
\details{
The \code{limits} matrix must have the following columns
\enumerate{
   \item targetColumn: A \code{character} value. Column on which restrictions should apply.
   \item min: A \code{numeric} value. Lower bound for the \code{targetColumn}. Should be set to \code{NA} when not applicable.
   \item max: A \code{numeric} value. Upper bound for the \code{targetColumn}. Should be set to \code{NA} when not applicable.
}
}
\section{Conversion rules (SQL -> R)}{

\enumerate{
   \item \code{enum} -> \code{factor}
   \item \code{int} -> \code{integer}
   \item \code{double} -> \code{numeric}
   \item \code{date} or \code{datetime} -> \code{Date}
   \item \code{varchar} -> \code{character}
}
}

