% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteBoundRestrictions.R
\name{pasteBoundRestrictions}
\alias{pasteBoundRestrictions}
\title{Paste inequality restrictions to a SQL query}
\usage{
pasteBoundRestrictions(query, targetVar, min, max, whereClause = TRUE)
}
\arguments{
\item{query}{Initial value of the query string. Restrictions are paster after it.}

\item{min}{Lower bound for the \code{targetVar}. Should be set to \code{NULL} when not applicable}

\item{max}{Upper bound for the \code{targetVar}. Should be set to \code{NULL} when not applicable}

\item{whereClause}{If set to \code{TRUE}, a 'where' clause is inserted before the restrictions
If set to \code{FALSE}, an 'and' clause is inserted}

\item{tagetVar}{Variable to be restricted}
}
\value{
The new query string
}
\description{
Concatenates inequality restrictions and an existing SQL query
}

