% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteLimitRestrictions.R
\name{pasteLimitRestrictions}
\alias{pasteLimitRestrictions}
\title{Paste inequality restrictions to a SQL query}
\usage{
pasteLimitRestrictions(query, targetVar, min, max, whereClause = TRUE)
}
\arguments{
\item{query}{A \code{character} value. Initial value of the query string. Restrictions will be pasted after it.}

\item{min}{A \code{numeric} value. Lower bound for the \code{targetVar}. Should be set to \code{NULL} when not applicable}

\item{max}{A \code{numeric} value. Upper bound for the \code{targetVar}. Should be set to \code{NULL} when not applicable}

\item{whereClause}{If set to \code{TRUE}, a 'where' clause is inserted before the restrictions
If set to \code{FALSE}, an 'and' clause is inserted}

\item{tagetVar}{A \code{character} value. Variable on which restrictions should apply.}
}
\value{
A \code{character}. The new query string.
}
\description{
Concatenates inequality restrictions and an existing SQL query string
}

