% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printChartData.R
\name{printChartData}
\alias{printChartData}
\title{Print chart data}
\usage{
printChartData(data, chart)
}
\arguments{
\item{data}{A \code{data.frame}. The data to be written.}

\item{chart}{A \code{character}. The name of the chart or of the file in which the chart was printed.}
}
\value{
A \code{character}. The name of the file that was printed, including its extension (.csv)
}
\description{
Writes chart data in a .csv file
}

