% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printGoogleChart.R
\name{printGoogleChart}
\alias{printGoogleChart}
\title{Print Google Chart}
\usage{
printGoogleChart(filename, chartObj)
}
\arguments{
\item{filename}{A \code{character} value. The name of the output file. The extension (.html) is not needed.}

\item{chartObj}{A \code{gvis} object.}
}
\value{
\code{TRUE} if the file was correctly created and
        \code{FALSE}, otherwise.
}
\description{
Prints a Google Chart to a .html file, encoded with UTF-8.
}
\seealso{
\code{\link[googleVis]{print.gvis}}
}

