% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateConsistency.R
\name{validateConsistency}
\alias{validateConsistency}
\title{Validate consistency between chart parameters}
\usage{
validateConsistency(data, timeVar, groupVar, targetVar)
}
\arguments{
\item{data}{A \code{data.frame}. The data source.}

\item{timeVar}{A \code{character}. The name of the column with the time variable.}

\item{groupVar}{A \code{character}. The name of the column with the group variable.}

\item{targetVar}{A \code{character vector}. The name(s) of the column(s) with the target variable(s) values.}
}
\value{
A \code{character}. The constant \code{.VALID} if variables are consistent and
        an error message, otherwise.
}
\description{
Checks if the chosen parameters follows the consistency rules.
             It uses a \code{data.frame} as the data source.
}
\section{Consistency rules}{

\enumerate{
   \item If \code{targetVar} has more than one element, \code{groupVar} must be \code{NULL}. Otherwise, \code{groupVar} must be a \code{factor} in R and \code{enum} in the database.
   \item If \code{groupVar} IS NOT \code{NULL}, each of its categories must have non null \code{targetVar} values corresponding to each different \code{timeVar} value. Besides, the set of different \code{timeVar} values must be the same for each category in \code{groupVar}.
   \item If \code{groupVar} IS \code{NULL}, each \code{targetVar} column must contain non null values corresponding to each different \code{timeVar} value. Besides, the set of different \code{timeVar} values must be the same for each \code{targetValue}.
}
}

