% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateGoogleChartParameters.R
\name{validateGoogleChartParameters}
\alias{validateGoogleChartParameters}
\title{Validate the parameters of a Google Chart}
\usage{
validateGoogleChartParameters(data, table, targetVar, groupVar, timeVar, min,
  max, restrictions)
}
\arguments{
\item{data}{A \code{data.frame}. The data source.}

\item{table}{A \code{character} value. The name of the table from which the data was retrieved.}

\item{targetVar}{A \code{character vector}. The name(s) of the column(s) with the target variable(s) values.}

\item{groupVar}{A \code{character}. The name of the column with the group variable.}

\item{timeVar}{A \code{character}. The name of the column with the time variable.}

\item{min}{A \code{numeric} value or a \code{date string} in the format 'yyyy-mm-dd'. Lower bound for the \code{timeVar}}

\item{max}{A \code{numeric} value or a \code{date string} in the format 'yyyy-mm-dd'. Upper bound for the \code{timeVar}}

\item{restrictions}{A n x 2 \code{matrix}. The n equality restrictions that make timeVar and groupVar values unique when combined.}
}
\value{
A \code{character}. The constant \code{.VALID} if the parameters are and
        an error message, otherwise.
}
\description{
Checks if the chosen parameters follows the validation rules for a Google Chart.
             It uses a \code{data.frame} as the data source.
}
\section{Validation rules}{

\enumerate{
   \item min, max, timeVar: \code{\link{validateLimits}}
   \item timeVar, groupVar, restrictions: \code{\link{validateKeys}}
   \item timeVar, groupVar, targetVar, restrictions: \code{\link{validateConsistency}}
}
}

