% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateGroupVariable.R
\name{validateGroupVariable}
\alias{validateGroupVariable}
\title{Validate Group Variable}
\usage{
validateGroupVariable(data, groupVar)
}
\arguments{
\item{data}{A \code{data.frame} with the actual data source or a \code{character} containing the name of the table in the dzVis database.
If it is a \code{data.frame}, a connection to the dzVis database WILL NOT be estabilished.
If it is a \code{character}, a connection WILL be estabilished.}

\item{groupVar}{A \code{character}. The name of the column to be checked.}
}
\value{
A \code{character}. The constant \code{.VALID} if the variable is valid and
        an error message, otherwise.
}
\description{
Checks the validity of a group variable, that is, a variable or set of variables that represents categories.
             It uses a direct connection to the dzVis database or a \code{data.frame} as the data source.
}
\section{Validation rules}{

The \code{groupVar} column type must be a \code{factor} in R and \code{enum} in the database.
}

