% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateGroupVariables.R
\name{validateGroupVariables}
\alias{validateGroupVariables}
\title{Validate Group Variables}
\usage{
validateGroupVariables(data, groupVar)
}
\arguments{
\item{data}{A \code{data.frame} with the actual data source or a \code{character} containing the name of the table in the dzVis database.
If it is a \code{data.frame}, a connection to the dzVis database WILL NOT be estabilished.
If it is a \code{character}, a connection WILL be estabilished.}

\item{groupVar}{A \code{character vector}. The name(s) of the column(s) to be checked.}
}
\value{
\code{TRUE} if \code{groupVar} is valid and
        \code{FALSE}, otherwise.
}
\description{
Checks the validity of a group variable, that is, checks whether the parameter variable or set of variables represents categories or not.
             It uses a direct connection to the dzVis database or a \code{data.frame} as the data source.
}
\section{Validation rules}{

\enumerate{
   \item If \code{groupVar} contains only one element, the \code{groupVar} column in \code{data} must be a \code{factor} in R and of type \code{enum} in the database.
   \item If \code{groupVar} contains more than one element, the corresponding columns in \code{data} must be \code{numeric} in R and of type \code{int} or \code{double} in the database.
}
}

