% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateKeys.R
\name{validateKeys}
\alias{validateKeys}
\title{Validate key parameters}
\usage{
validateKeys(table, candidateKeys, connection = NULL)
}
\arguments{
\item{table}{A \code{character}. The name of the table in the dzVis database.}

\item{keys}{A \code{character vector}. The column name(s) of the candidate key(s).}
}
\value{
A \code{character}. The constant \code{.VALID} if the parameters are valid.
        Otherwise, a \code{vector} containing the names of the columns that should be restricted to make \code{keys} values unique when combined.
}
\description{
Checks if the values of a parameter are unique within a data table.
             It uses a direct connection to the dzVis database.
}
\section{Validation rules}{

\enumerate{
   \item If \code{keys} has only one column name, each row of this column must contain a different value.
   \item If \code{keys} has more than one column name, a subset of those columns must have unique values when combined.
}
}

