% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateLimits.R
\name{validateLimits}
\alias{validateLimits}
\title{Validate the limits imposed on a column}
\usage{
validateLimits(data, columnName, min, max)
}
\arguments{
\item{data}{A \code{data.frame} with the actual data source or a \code{character} containing the name of the table in the dzVis database.
If it is a \code{data.frame}, a connection to the dzVis database WILL NOT be estabilished.
If it is a \code{character}, a connection WILL be estabilished.}

\item{columnName}{A \code{character}. The name of the column to be checked.}

\item{min}{A \code{numeric} value, a \code{character} or a \code{Date} in the format 'yyyy-mm-dd'. Lower bound for the column value. Should be set to \code{NULL} when not applicable.}

\item{max}{A \code{numeric} value, a \code{character} or a \code{Date} in the format 'yyyy-mm-dd'. Upper bound for the column value. Should be set to \code{NULL} when not applicable.}
}
\value{
A \code{character}. The constant \code{.VALID} if the limits are valid and
        an error message, otherwise.
}
\description{
Checks the validity of inequality restrictions imposed on a column.
             It uses a direct connection to the dzVis database or a \code{data.frame} as the data source.
}
\section{Validation rules}{

\enumerate{
   \item \code{min} and \code{max} must be of the same type: both \code{numeric} or both \code{character}.
   \item \code{min} must be less than \code{max}
   \item If \code{min} and \code{max} type is \code{numeric}, the column values must also be \code{numeric} in R and \code{int} or \code{double} in the database.
   \item If \code{min} and \code{max} type is \code{character}, the column values must be \code{character} in R and \code{varchar} in the database.
   \item If \code{min} and \code{max} type is \code{Date}, the column values must be \code{Date} in R and \code{date} or \code{datetime} in the database.
}
}

