% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateTargetVariables.R
\name{validateTargetVariables}
\alias{validateTargetVariables}
\title{Validate Target Variables}
\usage{
validateTargetVariables(data, targetVar)
}
\arguments{
\item{data}{A \code{data.frame} with the actual data source or a \code{character} containing the name of the table in the dzVis database.
If it is a \code{data.frame}, a connection to the dzVis database WILL NOT be estabilished.
If it is a \code{character}, a connection WILL be estabilished.}

\item{targetVar}{A \code{character vector}. The name(s) of the column(s) to be checked.}
}
\value{
A \code{character}. The constant \code{.VALID} if the variables are valid and
        an error message, otherwise.
}
\description{
Checks the validity of a target variable, a \code{numeric} variable usually of main interent.
             It uses a direct connection to the dzVis database or a \code{data.frame} as the data source.
}
\section{Validation rule}{

Each column whose name is in \code{targetVar} must be \code{numeric} in R and \code{double} or \code{int} in the database.
}

