% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\title{Connect to the dzVis database}
\usage{
connect()
}
\value{
The connection string to access the dzVis database
}
\description{
Connect to the dzVis database
}
\details{
Detailed information about the dzVis database can be found on
}
\seealso{
\code{\link{disconnect}}
}

