% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createComboChart.R
\name{createComboChart}
\alias{createComboChart}
\title{Create a Google Combo Chart}
\usage{
createComboChart(table, targetVar, groupVar, timeVar, min = NULL,
  max = NULL, restrictions = NULL)
}
\arguments{
\item{table}{A \code{character} value. The name of the table from which the data should be retrieved}

\item{targetVar}{A \code{character} value. The vertical axis variable}

\item{groupVar}{A \code{character} value. Variable by which \code{targetVar} should be grouped}

\item{timeVar}{A \code{character} value. The horizontal axis variable}

\item{min}{A \code{numeric} value or a \code{date string} in the format 'yyyy-mm-dd'. Lower bound for the \code{timeVar}}

\item{max}{A \code{numeric} value or a \code{date string} in the format 'yyyy-mm-dd'. Upper bound for the \code{timeVar}}

\item{restrictions}{n x 2 \code{matrix}. The n equality restrictions that make timeVar and groupVar unique when combined.}
}
\value{
\code{TRUE} if the chart was successfully created and
        \code{FALSE}, otherwise.
}
\description{
Creates a Combo Chart using the googleVis R package
}
\section{Parameters validation}{

The parameters are checked according to the rules estabilished in validation functions and in the following order:
\enumerate{
   \item targetVar: \code{\link{validateTargetVariable}}
   \item groupVar: \code{\link{validateGroupVariables}}
   \item timeVar: \code{\link{validateTimeVariable}}
   \item min, max, timeVar: \code{\link{validateLimits}}
   \item timeVar, groupVar, restrictions: \code{\link{validateKeys}}
   \item timeVar, groupVar: \code{\link{validateConsistency}}
}
}
\seealso{
\code{\link[googleVis]{gvisComboChart}}
}

