% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createComboChart.R
\name{createComboChart}
\alias{createComboChart}
\title{Create a Google Combo Chart}
\usage{
createComboChart(filename, table, targetVar, groupVar, timeVar, min = NA,
  max = NA, lineVar = NULL, operation = NULL, restrictions = NULL,
  alternatives = NULL)
}
\arguments{
\item{filename}{A \code{character}. The name of the .html file where the chart should be printed. The extension (.html) is not needed.}

\item{table}{A \code{character} value. The name of the table from which the data should be retrieved}

\item{targetVar}{A \code{character} value. The vertical axis variable}

\item{groupVar}{A \code{character} value. Variable by which \code{targetVar} should be grouped}

\item{timeVar}{A \code{character} value. The horizontal axis variable}

\item{min}{A \code{numeric} value or a \code{date string} in the format 'yyyy-mm-dd'. Lower bound for the \code{timeVar}}

\item{max}{A \code{numeric} value or a \code{date string} in the format 'yyyy-mm-dd'. Upper bound for the \code{timeVar}}

\item{restrictions}{A n x 2 \code{matrix}. The n equality restrictions that make timeVar and groupVar values unique when combined.}

\item{alternatives}{A n x 2 \code{matrix}. Alternative values of a column (joined by 'or' conditions in a SQL query). The first column must contain the names of the columns. The second, its values.}
}
\value{
A \code{character}. The constant \code{.SUCCESS} if the chart was successfully created.
        Otherwise, an error string returned by one of validation functions listed under the 'Parameters validation' section.
}
\description{
Creates a Combo Chart using the googleVis R package
}
\section{Parameters validation}{

The parameters are checked according to the rules estabilished in validation functions and in the following order:
\enumerate{
   \item targetVar, groupVar, timeVar, min, max, restrictions: \code{\link{validateGoogleChartParameters}}
   \item lineVar: \code{\link{validateTargetVariables}} and \code{\link{validateConsistency}}
}
}
\examples{
createComboChart("chart","pea_por_idade","percentual_total", "grupo_idade","mes","2010-01-01","2010-05-01", t(as.matrix(c("tipo_area","Total das áreas")))

}
\seealso{
\code{\link[googleVis]{gvisComboChart}}
}

