% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatDates.R
\name{formatDates}
\alias{formatDates}
\title{Format dates}
\usage{
formatDates(table, dates, toStandard = TRUE, connection = NULL)
}
\arguments{
\item{table}{xxxxxxxxxx}

\item{dates}{xxxxxxxxxxx}

\item{toStandard}{xxxxxxxxxx}

\item{connection}{xxxxxxx}
}
\value{
xxxxxxx
}
\description{
xxxxxxxxxx
}

