% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapChartVariables.R
\name{mapChartVariables}
\alias{mapChartVariables}
\title{Map chart variables}
\usage{
mapChartVariables(table, variables)
}
\arguments{
\item{table}{A \code{character} The name of a dzVis database table containing the variables - the columns - to be analysed.}

\item{variables}{A \code{character vector} The name(s) of the column(s).}
}
\value{
A \code{data.frame}. The first column represents the names of the variables. The next columns, named after the possible categories, booleans indicating if the variable can be assigned to the category.
}
\description{
Maps which categories chart variables can be assigned to. Minimum and maximum values are also provided when applicable.
}
\section{Possible categories}{

\enumerate{
   \item \code{TARGET} Numeric variables. They are typcally represented on the axis of the chart.
   \item \code{GROUP} Character variables. Represent categories by which the \code{TARGET} variables are classified.
   \item \code{TIME} Date variable. Indicates when the \code{TARGET} were observed.
   \item \code{KEY} Variables of any type. When combined, they represent the primary key - the id - of the table.
}
}

\section{Validation rules}{

The variables are checked according to the rules estabilished in validation functions and in the following order:
\enumerate{
   \item targetVar: \code{\link{validateTargetVariables}}
   \item groupVar: \code{\link{validateGroupVariable}}
   \item timeVar: \code{\link{validateTimeVariable}}
   \item key: \code{\link{validateKeys}}
}
}

