% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteIdRestrictions.R
\name{pasteIdRestrictions}
\alias{pasteIdRestrictions}
\title{Paste equality restrictions to a SQL query}
\usage{
pasteIdRestrictions(query, restrictions, whereClause = TRUE)
}
\arguments{
\item{query}{A \code{character} value. Initial value of the query string. Restrictions will be pasted after it.}

\item{restrictions}{A n x 2 \code{matrix} with n equality restrictions}

\item{whereClause}{If set to \code{TRUE}, a 'where' clause is inserted before the restrictions.
If set ot \code{FALSE}, an 'and' clause is inserted}
}
\value{
A \code{character}. The new query string.
}
\description{
Concatenates equality restrictions and an existing SQL query string
}

