% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteIdRestrictions.R
\name{pasteIdRestrictions}
\alias{pasteIdRestrictions}
\title{Paste equality restrictions to a SQL query}
\usage{
pasteIdRestrictions(query, restrictions, whereClause = TRUE)
}
\arguments{
\item{query}{Initial value of the query string. Restrictions are pasted after it.}

\item{restrictions}{n x 2 matrix with n equality restrictions}

\item{whereClause}{If set to \code{TRUE}, a 'where' clause is inserted before the restrictions.
If set ot \code{FALSE}, an 'and' clause is inserted}
}
\value{
The new query string
}
\description{
Concatenates equality restrictions and an existing SQL query
}

