% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteLimitRestrictions.R
\name{pasteLimitRestrictions}
\alias{pasteLimitRestrictions}
\title{Paste inequality restrictions to a SQL query}
\usage{
pasteLimitRestrictions(query, restrictions, whereClause = FALSE)
}
\arguments{
\item{query}{A \code{character} value. Initial value of the query string. Restrictions will be pasted after it.}

\item{restrictions}{A n x 3 \code{matrix} with n inequality restrictions. See the details section for important information.}

\item{whereClause}{If set to \code{TRUE}, a 'where' clause is inserted before the restrictions
If set to \code{FALSE}, an 'and' clause is inserted}
}
\value{
A \code{character}. The new query string.
}
\description{
Concatenates inequality restrictions and an existing SQL query string
}
\details{
The \code{restrictions} matrix must have the following columns:
\enumerate{
   \item targetColumn: A \code{character} value. Column on which restrictions should apply.
   \item min: A \code{numeric} value. Lower bound for the \code{targetColumn}. Should be set to \code{NA} when not applicable.
   \item max: A \code{numeric} value. Upper bound for the \code{targetColumn}. Should be set to \code{NA} when not applicable.
}
}

