% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printGoogleChart.R
\name{printGoogleChart}
\alias{printGoogleChart}
\title{Print Google Chart}
\usage{
printGoogleChart(chartObj, filename = NULL)
}
\arguments{
\item{chartObj}{A \code{gvis} object.}

\item{filename}{A \code{character} value. The name of the output file. The extension (.html) is not needed.}
}
\value{
\code{character}. The name of the file where the chart was printed.
}
\description{
Prints a Google Chart to a .html file, encoded with UTF-8.
}
\seealso{
\code{\link[googleVis]{print.gvis}}
}

